+++ oshi-common/src/main/java/oshi/software/common/os/mac/MacNetworkParams.java
@@ -12,6 +12,7 @@ import oshi.software.os.NetworkParams;
 import oshi.util.ExecutingCommand;
 import oshi.util.ParseUtil;
 import oshi.util.driver.unix.NetstatRoute;
+import oshi.util.driver.unix.RouteTableDump;
 
 /**
  * Common Mac NetworkParams logic shared between the JNA and FFM implementations. The default-gateway lookups are
@@ -55,7 +56,25 @@ public abstract class MacNetworkParams extends AbstractNetworkParams {
 
     @Override
     public List<NetworkParams.IPRoute> getRoutes() {
+        byte[] dump = queryRouteDump();
+        if (dump.length > 0) {
+            List<NetworkParams.IPRoute> routes = RouteTableDump.parse(dump, RouteTableDump.MACOS,
+                    queryInterfaceNameByIndex());
+            if (!routes.isEmpty()) {
+                return routes;
+            }
+        }
         return NetstatRoute.queryRoutes("netstat -rn -f inet", "netstat -rn -f inet6", IF_NAME_INDEX,
                 queryInterfaceIndexByName());
+    }
+
+    /**
+     * Fetches the kernel's routing table dump.
+     *
+     * @return The bytes of a {@code NET_RT_DUMP} sysctl, or an empty array if this build cannot make the call, in which
+     *         case the routing table is read by running a command instead
+     */
+    protected byte[] queryRouteDump() {
+        return new byte[0];
     }
 }